+++ Assets/UI/Scripts/EntityHpBarUI.cs
@@ -14,9 +14,10 @@ namespace PipelineDreams {
                     var s = HPBarFull.rect; HPBar.sizeDelta = new Vector2(s.width * v, 0);
                     HPBarBackground.sizeDelta = new Vector2(s.width * (1 - v), 0); HPBar.gameObject.SetActive(v != 1);
                 };
-                esb.entity.GetComponent<EntityHealth>().OnDamagedAmount += (d, h, e) => {
+                esb.entity.GetComponent<EntityHealth>().OnDamagedAmount += (d, h, e) => 
+                {
                     var obj = Instantiate(DT, transform, false);
-                    obj.Init(d);
+                    obj.Init((int)d);
                     obj.transform.localPosition = new Vector3(0, 100, 0);
 
                 };
